backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
pandas==2.1.3
scikit-learn==1.3.2
catboost==1.2.2
numpy==1.25.2
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
